Estimate N on the basis of two pilot samples.

Estimation as described by Chakraborti, Hong, & van de Wiel (2006).

n_locshift(s1, s2, delta, alpha = 0.05, power = 0.9, q = 0.5)

Arguments

s1, s2

pilot samples

delta

numeric value, location shift parameter \(\delta\)

alpha

type-I error probability

power

1 - type-II error probability, the desired statistical power

q

size of group0 relative to total sample size.

Value

Returns an object of class "sample_size". It contains the following components:

N

the total sample size

n0

sample size in Group 0 (control group)

n1

sample size in Group 1 (treatment group)

two_sided

logical, TRUE, if the estimated sample size refers to a two-sided test

alpha

type I error rate used in sample size estimation

power

target power used in sample size estimation

effect

effect size used in sample size estimation

effect_type

short description of the type of effect size

comment

additional comment, if there is any

call

the matched call.

Details

WARNING: Note that the estimation has high variability due to its dependence on pilot samples. The smaller the pilot sample, the more uncertain is the estimation of the required sample size. In a simulation study, we found that the method may also be inaccurate on average, depending on the investigated data.

References

Chakraborti, S., Hong, B., & van de Wiel, M. A. (2006). A note on sample size determination for a nonparametric test of location. Technometrics, 48(1), 88–94. https://doi.org/10.1198/004017005000000193

Examples

n_locshift(s1 = rexp(10), s2 = rexp(10),
           alpha = 0.05, power = 0.9, delta = 0.35)
#> Estimated sample size for group difference.
#> Wilcoxon-Mann-Whitney Test, Location shift 
#> 
#> N (total)		 123.47 
#> n0 (Group 0)		 61.74 
#> n1 (Group 1)		 61.74 
#> 
#> Effect size		 0.35 
#> Effect type		 location shift 
#> Type I error		 0.05 
#> Target power		 0.9 
#> Two-sided		 FALSE 
#> 
#> Call: n_locshift(s1 = rexp(10), s2 = rexp(10), delta = 0.35, alpha = 0.05, 
#>     power = 0.9)