Compute a distribution of estimates of N based on two pilot samples.

Estimation of sample sizes based on resampled pilot samples from the empirical cumulative density. Based on the work of Chakraborti, Hong, & van de Wiel (2006).

resample_n_locshift(
  s1,
  s2,
  delta,
  alpha = 0.05,
  power = 0.9,
  n_resamples = 500,
  q = 0.5
)

Arguments

s1, s2

Pilot samples

delta

numeric value, location shift parameter \(\delta\)

alpha

Type I error probability

power

1 - Type II error probability, the desired statistical power

n_resamples

number of resamples to use in bootstrapping

q

size of group0 relative to total sample size.

Value

numeric vector of sample size estimates (total sample size)

Details

WARNING: Note that the estimation has high variability due to its dependence on pilot samples. The smaller the pilot sample, the more uncertain is the estimation of the required sample size. In a simulation study, we found that the method may also be inaccurate on average, depending on the investigated data.

References

Chakraborti, S., Hong, B., & van de Wiel, M. A. (2006). A note on sample size determination for a nonparametric test of location. Technometrics, 48(1), 88–94. https://doi.org/10.1198/004017005000000193